AVS returns HTTP/2 Multipart messages to it clients. See https://developer.amazon.com/docs/alexa-voice-service/structure-http2-request.html for more details on the exact structure.
Added a parser that takes the multi-part reponse and extracts the 'caption' which represents the actual content that should be displayed to the user as Alexa's response.

The rest of the message contains audio bytes of the same response, SSML tagsetc., all of which gets dropped.

We are using https://github.com/miguelmota/http-message-parser/ to do the parsing.

Unit tests added.

Note: http-message-parser has a bug in the latest version and so package.json is configured to not use the latest version of this package but use 0.0.22.
A bug has been filed against the developer at https://github.com/miguelmota/http-message-parser/issues/4