Preliminary integration of ChatWindow with AVSGateway. We still haven't plumbed through the real access tokens and so for now, the calls to AVSGateway will always return a 403 and as a result, Alexa will effectively respond with an error message all the time.
Having said that, this is a good step forward because it basically
makes the chatwindow interactive for the first time. User can type
something and get a response from Alexa, even though it is just an
error message each time.

In a later commit, we just need to plumb the access token through.

There was a different commit that made some formatting related settings
changes and as a result, this commit picked up some format changes.