Introduced AVSGateway class to interact with Alexa (#12)
* Introduced AVSGateway class.

This class makes HTTP requests to AVS and parses the responses.
The class is currently tailor made just for the Text::TextMessage API
on AVS and supports a method that returns Alexa's response for a
given request. It might be turned into a full fledged gateway that supports
all APIs on AVS on a future date.

Tests written.

Since we need to render custom error messages in response to different
kinds of errors, this commit also introduces additinal modules to
map error codes to human readable messages.