Pass real access token to AVSGateway.
Now that the access_token has been plumbed through the state, in this commit we start using it by passing the real access_token to AVSGateway. We have an end-end functioning silent Alexa with this commit.

Note that for now, we blindly pass the access_token regardless of whether or not it is valid and if it is even present. The assumption is that it is either

the 'App' component's responsibility to kick the user back to a login screen whenever the access_token is missing or expired.
OR
AVSGateway's responsibility to notify App to kick the user out whenever it receives a 403 from Alexa or notices a missing access_token.
Said logout mechanism will be implemented in a later commit.