Custom styles for chat bubbles.
Added colors to the char bubbles. The colors aren't final but this
commit is just about getting the code in place to be able to easily
adjust colors.

As of now, react-chat-ui has a bug that prevents the styles from
getting applied properly.
I submitted a pull request to react-chat-ui to fix the bug
https://github.com/brandonmowat/react-chat-ui/pull/26

I tested with the bug fix locally and the styles apply correctly.

Having said that, the bug only prevents the CSS from applying properly
and doesn't break the app or tests and so it should be safe to review
and commit this code.