Fix for https://github.com/s-maheshbabu/silent-alexa/issues/22
Alexa can return her response in multiple parts. Previously, we
assumed that the parts are text and audio responses. However, as
noted in the above issue, the text itself can be broken into
multiple parts.

In the current code, since we assume that there is only one text
part, we end up showing only the first part in her multi-part
response message.

This is a preparatory commit that changes the interfaces between
ChatWindow / AVSGateway / SpeakDirectiveParser to be list of responses
instead of a single response string. There is no behavior change in
this commit besides the interface change.