Now the code base has grown, a flat structure is becoming unweildy.
This is the first stab at organizing the code into a
hierarchical structure. There is no right way to organize
a React project and so this is something that has to
constantly evolve as we add more pieces.

Note: Containers (as against everything being a component)
appears to be a fairly standard practice and something we
should refactor out soon. To keep this CR simple, I left the
components as is and just moved them into their own directory.