Alexa sends an 'ExpectSpeech' directive along with a 'Speak' directive for multi-turn dialogs. The intent is to tell the client to re-open the microphone stream.
Since Silent Alexa is a text based app, 'ExpectSpeech' directive
is not really relevant (may be we can show a visual indicator
that Alexa is expecting the user to say something but it is an
immediate requirement).

Instead of failing on unexpected directives, this change is to
ignore any directive other than the 'Speak' directive.