Changing homepage URL
Having a path /silent-alexa/ is resulting in some custom routing logic to be written. Since eventually this app is envisioned to be a single page app, there is no point writing all the custom routing logic. So, changing the homepage in the dev environment to not contain any paths to make out lives easier.