Switch LWA token management to use cookies. (#63)
In this commit, we are switching to use cookies to manage LWA tokens. The commit unfortunately spiraled into a pretty large commit.

Changing the authentication management to use cookies made the previous method of component drilling quite cumbersome and so we ended up switching to use Router Context concept. This in turn resulted in having to get rid of the previous AuthenticationInfo class which in turn resulted tin having to touch a lot of files.

There were some changes made to routing as well because of the change in authentication info component structure.

At this point, we are able to have a functional app with authentication information stored in cookies and authentication information passed across the app using Context API. All the tests were updated.

Finally, a bunch of dependencies were also updated.