Redirect user to a login page if session expired.
When the user session expires (authentication token issues by Amazon expires),
we need to redirect the user away from the protected pages onto a public
page. In this commit, we redirect the user to /access_denied but for the
time being. We might want to just route them to home page.