Introduce integration tests.
A recent bug proved that unit testing is not adequate. The bug was the
logout button not working as intended. The button would change to
'Login' and clear access tokens which is what we unit test for. However,
because of a bug in a different part in the code, the chat window was
not disappearing.

This underscored the need for integration tests. react-testing-library
is a strong proponent of integration tests and provides many utility
functions to effectively write integration tests.

In this commit, we write a single integration test covering the above described
scenario using react-testing-library. The idea is to continue adding
more integration tests going forward.