Introduce loading animation.
We are now making a background call to Amazon to send AddOrUpdateReport
event. According to AVS, users should not be able to interact until
after this event is successfully posted. This commit is to introduce a
loading animation to be shown to the user while the background calls are
being made.