Refactor to declarative routing.
There was a bug that is resulting in both welcome screen and chat
window being rendered. While there is probably a solution for that
even in the old static routing, I upgraded to use the newer and
recommended declarative routing. The new style also makes is very
straight forward to fix the bug.